Alert

+++ components/FavoritesComponent.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
@@ -28,11 +28,28 @@ class Favorites extends Component {
     const renderFavoriteItem = ({ item }) => {
       const rightButton = [
         {
-          text: 'Delete',
-          type: 'delete',
-          onPress: () => this.props.deleteFavorite(item.id)
+            text: 'Delete', 
+            type: 'delete',
+            onPress: () => {
+                Alert.alert(
+                    'Delete Favorite?',
+                    'Are you sure you wish to delete the favorite campsite ' + item.name + '?',
+                    [
+                        { 
+                            text: 'Cancel', 
+                            onPress: () => console.log(item.name + 'Not Deleted'),
+                            style: ' cancel'
+                        },
+                        {
+                            text: 'OK',
+                            onPress: () => this.props.deleteFavorite(item.id)
+                        }
+                    ],
+                    { cancelable: false }
+                );
+            }
         }
-      ];
+    ];
 
       return (
         <Swipeout right={rightButton} autoClose={true}>
